Added game files,
I have added all of the original files from the game as well as the edits to background images intro 1 and intro 2 to remove text. These images now have a blank text box ready to be overlayed with live editable text.
I have updated the license to reflect the changes that we plan to make to the project in accordance with the terms of the license provided by the developers.